把主要内容从__init__移入main并用black格式化了代码，其余一点都没改
